main.py
import cv2
import numpy as np
import io, base64, time, os, jwt
from fastapi import FastAPI, UploadFile, File, HTTPException, Depends, status
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
from PIL import Image
from typing import List, Dict, Any

# --- CONFIGURACIÓN E INICIALIZACIÓN ---
app = FastAPI(title="AgroVision API - ITM 1")
SECRET_KEY = os.environ.get('JWT_SECRET', 'AgroVisionKeyParaITM1') 
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/auth/login")

# ... (El código de modelos, USERS, TRACE_LOGS, log_trace, y decode_token va aquí) ...

# --- ENDPOINTS OBLIGATORIOS Y CORE ---

@app.get("/health")
def health_check():
    """Endpoint de Diagnóstico Básico requerido por el Portal Docente"""
    # CRÍTICO: El profesor verifica que este endpoint responda OK [cite: 168]
    return {"status": "ok", "service": "AgroVision Backend ITM1"}

@app.post("/api/auth/login")
def login(user: UserIn):
    # ... (Lógica de login con JWT) ...
    
@app.post("/api/diagnose")
async def diagnose_image(file: UploadFile = File(...), current_user: dict = Depends(get_current_user)):
    # ... (Lógica de Segmentación HSV [cite: 87] y clasificación CNN [cite: 89] va aquí) ...
    # ... (El registro en la Trazabilidad [cite: 140] también va aquí) ...